+++ R/balance_ions.R
@@ -1,19 +1,28 @@
 #' @title Add an ion to balance overall charge in a water
 #'
-#' @description This function takes a water defined by \code{\link{define_water}} and balances charge.
+#' @description This function takes a water defined by [define_water] and balances charge.
+#' For a single water use `balance_ions`; for a dataframe use `balance_ions_chain`.
+#' Use [pluck_water] to get values from the output water as new dataframe columns.
 #'
-#' @details If more cations are needed, a default of sodium will be added. User may specify which cation ("na", "k", "ca", or "mg") to use for balancing.
+#' @details If more cations are needed, sodium will be added. User may specify which cation ("na", "k", "ca", or "mg") to use for balancing.
 #' If calcium and magnesium are not specified when defining a water with
-#' \code{\link{define_water}}, they will default to 0 and not be changed by this function unless specified in the cation argument.
-#' Anions are added by default with chloride. User may specify which anion ("cl", "so4") to use for balancing. This function is purely mathematical.
+#' [define_water], they will default to 0 and not be changed by this function unless specified in the cation argument.
+#' Anions are added by default with chloride. User may specify which anion ("cl", "so4") to use for balancing.
+#' This function is purely mathematical.
 #' User should always check the outputs to make sure values are reasonable for the input source water.
 #'
-#' @param water Water created with define_water, which may have some ions set to 0 when unknown
+#' For large datasets, using `fn_once` or `fn_chain` may take many minutes to run. These types of functions use the furrr package
+#'  for the option to use parallel processing and speed things up. To initialize parallel processing, use
+#'  `plan(multisession)` or `plan(multicore)` (depending on your operating system) prior to your piped code with the
+#'  `fn_once` or `fn_chain` functions. Note, parallel processing is best used when your code block takes more than a minute to run,
+#'  shorter run times will not benefit from parallel processing.
+#'
+#' @param water Water created with [define_water], which may have some ions set to 0 when unknown
 #' @param anion Selected anion to use to for ion balance when more cations are present. Defaults to "cl". Choose one of c("cl", "so4").
 #' @param cation Selected cation to use to for ion balance when more anions are present. Defaults to "na". Choose one of c("na", "k", "ca", or "mg").
 #'
 #' @examples
-#' water_defined <- define_water(7, 20, 50, 100, 80, 10, 10, 10, 10, tot_po4 = 1) %>%
+#' water_defined <- define_water(7, 20, 50, 100, 80, 10, 10, 10, 10) %>%
 #'   balance_ions()
 #'
 #' water_defined <- define_water(7, 20, 50, tot_hard = 150) %>%
@@ -21,7 +30,7 @@
 #'
 #' @export
 #'
-#' @returns A water class object with updated ions to balance water charge.
+#' @returns `balance_ions` returns a single water class object with updated ions to balance water charge.
 #'
 balance_ions <- function(water, anion = "cl", cation = "na") {
   if (!methods::is(water, "water")) {
@@ -79,11 +88,11 @@ balance_ions <- function(water, anion = "cl", cation = "na") {
       add_k <- add_cat
       k_new <- ifelse(is.na(water@k), add_k, water@k + add_k)
       water@estimated <- paste(water@estimated, "k", sep = "_")
-    }else if (cation == "ca") {
+    } else if (cation == "ca") {
       add_ca <- add_cat / 2
       ca_new <- ifelse(is.na(water@ca), add_ca, water@ca + add_ca)
       water@estimated <- paste(water@estimated, "ca", sep = "_")
-    }else if (cation == "mg") {
+    } else if (cation == "mg") {
       add_mg <- add_cat / 2
       mg_new <- ifelse(is.na(water@mg), add_mg, water@mg + add_mg)
       water@estimated <- paste(water@estimated, "mg", sep = "_")
@@ -118,7 +127,7 @@ balance_ions <- function(water, anion = "cl", cation = "na") {
   if (grepl("tds", water@estimated) & grepl("cond", water@estimated)) {
     # Update TDS and cond if they were estimated from IS. Otherwise, assume initial values were measured.
     water@tds <- water@tds + convert_units(add_na, "na", "M", "mg/L") + convert_units(add_k, "k", "M", "mg/L") +
-      convert_units(add_ca, "ca", "M", "mg/L") +  convert_units(add_mg, "mg", "M", "mg/L") +
+      convert_units(add_ca, "ca", "M", "mg/L") + convert_units(add_mg, "mg", "M", "mg/L") +
 
       convert_units(add_cl, "cl", "M", "mg/L") + convert_units(add_so4, "so4", "M", "mg/L")
 
@@ -130,84 +139,11 @@ balance_ions <- function(water, anion = "cl", cation = "na") {
   return(water)
 }
 
-#' Apply `balance_ions` function and output a dataframe
+#' @rdname balance_ions
 #'
-#' This function allows \code{\link{balance_ions}} to be added to a piped data frame.
-#' tidywater functions cannot be added after this function because they require a `water` class input.
-#'
-#'  For large datasets, using `fn_once` or `fn_chain` may take many minutes to run. These types of functions use the furrr package
-#'  for the option to use parallel processing and speed things up. To initialize parallel processing, use
-#'  `plan(multisession)` or `plan(multicore)` (depending on your operating system) prior to your piped code with the
-#'  `fn_once` or `fn_chain` functions. Note, parallel processing is best used when your code block takes more than a minute to run,
-#'  shorter run times will not benefit from parallel processing.
-#'
-#' @param df a data frame containing a water class column, which has already been computed using \code{\link{define_water_chain}}
+#' @param df a data frame containing a water class column, which has already been computed using [define_water_chain]
 #' @param input_water name of the column of water class data to be used as the input for this function. Default is "defined_water".
-#' @param anion Selected anion to use to for ion balance when more cations are present. Defaults to "cl". Choose one of c("cl", "so4").
-#' @param cation Selected cation to use to for ion balance when more anions are present. Defaults to "na". Choose one of c("na", "k", "ca", or "mg").
-#' @seealso \code{\link{balance_ions}}
-#'
-#' @examples
-#' library(purrr)
-#' library(furrr)
-#' library(tidyr)
-#' library(dplyr)
-#'
-#' example_df <- water_df %>%
-#'   define_water_chain() %>%
-#'   balance_ions_once(anion = "so4", cation = "mg")
-#'
-#' example_df <- water_df %>%
-#'   define_water_chain(output_water = "Different_defined_water_column") %>%
-#'   balance_ions_once(input_water = "Different_defined_water_column")
-#'
-#' # Initialize parallel processing
-#' plan(multisession, workers = 2) # Remove the workers argument to use all available compute
-#' example_df <- water_df %>%
-#'   define_water_chain() %>%
-#'   balance_ions_once()
-#'
-#' # Optional: explicitly close multisession processing
-#' plan(sequential)
-#'
-#' @import dplyr
-#' @importFrom tidyr unnest_wider
-#' @export
-#' @returns A dataframe with updated ions to balance water charge
-
-balance_ions_once <- function(df, input_water = "defined_water",
-                              anion = "cl", cation = "na") {
-  balance_df <- balanced_water <- NULL # Quiet RCMD check global variable note
-  output <- df %>%
-    mutate(balanced_water = furrr::future_pmap(
-      list(
-        water = !!as.name(input_water),
-        anion = anion,
-        cation = cation
-        ), balance_ions)
-      ) %>%
-    mutate(balance_df = furrr::future_map(balanced_water, convert_water)) %>%
-    unnest_wider(balance_df) %>%
-    select(-balanced_water)
-}
-
-#' Apply `balance_ions` within a dataframe and output a column of `water` class to be chained to other tidywater functions
-#'
-#' This function allows \code{\link{balance_ions}} to be added to a piped data frame.
-#' Its output is a `water` class, and can therefore be used with "downstream" tidywater functions.
-#'
-#'  For large datasets, using `fn_once` or `fn_chain` may take many minutes to run. These types of functions use the furrr package
-#'  for the option to use parallel processing and speed things up. To initialize parallel processing, use
-#'  `plan(multisession)` or `plan(multicore)` (depending on your operating system) prior to your piped code with the
-#'  `fn_once` or `fn_chain` functions. Note, parallel processing is best used when your code block takes more than a minute to run,
-#'  shorter run times will not benefit from parallel processing.
-#'
-#' @param df a data frame containing a water class column, which has already been computed using \code{\link{define_water_chain}}
-#' @param input_water name of the column of water class data to be used as the input for this function. Default is "defined_water".
-#' @param output_water name of the output column storing updated parameters with the class, water. Default is "balanced_water".
-#' @param anion Selected anion to use to for ion balance when more cations are present. Defaults to "cl". Choose one of c("cl", "so4").
-#' @param cation Selected cation to use to for ion balance when more anions are present. Defaults to "na". Choose one of c("na", "k", "ca", or "mg").
-#' @seealso \code{\link{balance_ions}}
+#' @param output_water name of the output column storing updated water classes. Default is "balanced_water".
 #'
 #' @examples
 #' library(purrr)
@@ -237,16 +173,17 @@ balance_ions_once <- function(df, input_water = "defined_water",
 #'
 #' @import dplyr
 #' @export
-#' @returns A data frame containing a water class column with updated ions to balance water charge.
+#' @returns `balance_ions_chain` returns a dataframe with a new column with the ion balanced water
 
 balance_ions_chain <- function(df, input_water = "defined_water", output_water = "balanced_water",
                                anion = "cl", cation = "na") {
+  validate_water_helpers(df, input_water)
   output <- df %>%
     mutate(!!output_water := furrr::future_pmap(
       list(
         water = !!as.name(input_water),
         anion = anion,
         cation = cation
-      ), balance_ions)
-    )
+      ), balance_ions
+    ))
 }