+++ NEWS.md
@@ -3,12 +3,17 @@
 * `calculate_corrosion` expanded CCPP search range (fewer errors)
 
 ## Breaking changes
+* Helper function (`_chain` and `_once`) behavior change: can now specify column name unquoted (eg, alum = AlumDose)
+* Removed most (`_once`) helper functions because the desired output is almost never all water slots as columns. 
+Refer to `pluck_water(parameter = "all")` for same behavior.
 * `pac_toc` now constrained to accept a smaller range of doses and times. The form of the equation was allowing negative TOC outputs.
 * `define_water`, `chemdose_ph`, `blend_waters`, and other pH related functions slightly impacted by a fix in concentration vs activity. 
 Previous code assumed pH = 10^-[H+], code has been corrected to pH = 10^-{H+}
 
 ## New features
 * User may now choose which cation or anion to use for balancing ions
+* dic now calculated in `define_water`
+* dic now available in `convert_units`
 
 # tidywater 0.7.0